feat: 담고 빼기 토글 기능 분리
- 증가, 감소 기능 분리하여 버튼에 별개로 작동하도록 수정

+++ src/hooks/useCartItems.tsx
@@ -1,4 +1,4 @@
-import { useCallback, useEffect, useMemo, useState } from "react";
+import { useEffect, useMemo, useState } from "react";
 import { CartItems } from "../types/cartItems";
 import { CartItemsAPI } from "../apis/cartItems";
 import { isErrorResponse } from "../utils/typeGuard";
@@ -7,8 +7,9 @@ type Props = React.Dispatch<React.SetStateAction<string>>;
 
 interface UseCartItemsReturn {
   cartItemsCount: number;
-  isProductInCart: (id: number) => boolean;
-  handleCartItemToggle: (productId: number) => Promise<void>;
+  decreaseItemQuantity: (productId: number) => Promise<void>;
+  increaseItemQuantity: (productId: number) => Promise<void>;
+  quantityByProductId: (productId: number) => number | undefined;
 }
 
 const useCartItems = (setErrorMessage: Props): UseCartItemsReturn => {
@@ -37,22 +38,36 @@ const useCartItems = (setErrorMessage: Props): UseCartItemsReturn => {
   );
 
   const cartItemsCount = cartItems?.content.length ?? 0;
+  const allQuantities = cartItems?.content.map(({ id, quantity }) => ({
+    productId: id,
+    quantity: quantity,
+  }));
+  const quantityByProductId = (productId: number) =>
+    allQuantities?.find((item) => item.productId === productId)?.quantity;
 
-  const isProductInCart = useCallback(
-    (id: number) => cartItemIds.some((item) => item.productId === id),
-    [cartItemIds]
-  );
-
-  const handleCartItemToggle = async (productId: number) => {
+  const decreaseItemQuantity = async (productId: number) => {
     const currentProductId = cartItemIds.find(
       (productInfo) => productInfo.productId === productId
     );
 
-    if (currentProductId) {
-      await CartItemsAPI.delete(currentProductId.cartId);
-    } else {
-      await CartItemsAPI.post(productId);
+    if (currentProductId) await CartItemsAPI.delete(currentProductId.cartId);
+
+    const response = await CartItemsAPI.get();
+
+    if (isErrorResponse(response)) {
+      setErrorMessage(response.error);
+      return;
     }
+
+    setCartItems(response as CartItems);
+  };
+
+  const increaseItemQuantity = async (productId: number) => {
+    const currentProductId = cartItemIds.find(
+      (productInfo) => productInfo.productId === productId
+    );
+
+    if (currentProductId) await CartItemsAPI.post(productId);
 
     const response = await CartItemsAPI.get();
 
@@ -66,8 +81,9 @@ const useCartItems = (setErrorMessage: Props): UseCartItemsReturn => {
 
   return {
     cartItemsCount,
-    isProductInCart,
-    handleCartItemToggle,
+    decreaseItemQuantity,
+    increaseItemQuantity,
+    quantityByProductId,
   };
 };
